Making started/finished event types in script optional
We don't want to track every time a repeated script is started or finished,
so we'll allow these columns to be null and only set them when we're creating
a repeating script.

+++ api/service/script/delete.class.php
@@ -36,8 +36,8 @@ class delete extends \cenozo\service\delete
     try
     {
       // delete the associated event types
-      $this->db_started_event_type->delete();
-      $this->db_finished_event_type->delete();
+      if( !is_null( $this->db_started_event_type ) ) $this->db_started_event_type->delete();
+      if( !is_null( $this->db_finished_event_type ) ) $this->db_finished_event_type->delete();
     }
     catch( \cenozo\exception\notice $e )
     {
